Split into handlers.
